requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
langchain-core==0.3.12
langchain-groq==0.2.0
langchain-community==0.3.2